treewide: split commands into separate files
