+++ src/connections.rs
@@ -1,0 +1,65 @@
+// Copyright 2022 Redpanda Data, Inc.
+//
+// Use of this software is governed by the Business Source License
+// included in the file BSL.md
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0
+
+use std::sync::Arc;
+use std::sync::Mutex;
+
+use tokio;
+use tokio::io::AsyncWriteExt;
+use tokio::net::TcpStream;
+
+use log::*;
+
+/// Stress the tcp_server_listen_backlog setting
+/// A system with a small backlog will experience errors here: a system with
+/// a larger backlog will not.
+pub async fn connections(addr_str: String, n: i32) {
+    let error_count = Arc::new(Mutex::new(0 as u32));
+    let mut tasks = vec![];
+    for _i in 0..n {
+        let addr_str_copy = addr_str.clone();
+        let error_count_ref = error_count.clone();
+        tasks.push(tokio::spawn(async move {
+            let con = TcpStream::connect(addr_str_copy).await;
+            match con {
+                Err(e) => {
+                    let mut locked = error_count_ref.lock().unwrap();
+                    *locked += 1;
+                    println!("Connection error: {}", e);
+                }
+                Ok(mut sock) => {
+                    for _j in 0..1000 {
+                        match sock.write_all("ohaihowyoudoing".as_bytes()).await {
+                            Ok(_bytes) => {}
+                            Err(e) => {
+                                println!("write error {}", e);
+                                let mut locked = error_count_ref.lock().unwrap();
+                                *locked += 1;
+                                break;
+                            }
+                        }
+                        sock.flush().await.unwrap();
+                    }
+                }
+            }
+        }));
+    }
+
+    for t in tasks {
+        t.await.unwrap();
+    }
+
+    let final_errs = *(error_count.lock().unwrap());
+    if final_errs > 0 {
+        error!("{} connection attempts failed", final_errs);
+        std::process::exit(-1);
+    } else {
+        info!("Completed with no connection errors.")
+    }
+}
